skelToSurfaceMesh produces spheres

+++ Mesh_Utilities/mesh_types.h
@@ -8,7 +8,7 @@
 #include <CGAL/boost/graph/split_graph_into_polylines.h>
 #include <CGAL/Gmpzf.h>
 
-typedef CGAL::Simple_cartesian<long double>                                              Kernel;
+typedef CGAL::Simple_cartesian<double>                                              Kernel;
 typedef Kernel::Point_3                                                             Point;
 typedef Kernel::Vector_3                                                            Vector;
 typedef Kernel::Segment_3                                                           Segment;
@@ -28,42 +28,6 @@ typedef Mean_Curvature_Flow_Skeleton::vertex_iterator                           
 typedef std::map<face_descriptor, double> Facet_double_map;
 
 typedef CGAL::Surface_mesh_deformation<Point3Mesh>                                  Point3Mesh_deformation;
-////typedef boost::property_map<Point3Mesh, CGAL::vertex_point_t>::type  VertexPointMap;
-//typedef boost::property_map<Point3Mesh, CGAL::vertex_point_t>::type  Vertex_Property_Map;
-//typedef Point3Mesh::template Property_map<vertex_descriptor, Vector> Vertex_Normal_Map;
-//
-//struct Bottom
-//{
-//    Bottom(Vertex_Property_Map pmap, Vertex_Normal_Map nmap, double vlen)
-//        : pmap(pmap), nmap(nmap), vlen(vlen)
-//    {}
-//    void operator()(const vertex_descriptor& vin, const vertex_descriptor vout) const
-//    {
-//        boost::put(pmap, vout, boost::get(pmap, vout) - vlen * boost::get(nmap, vin));
-//    }
-//    Vertex_Property_Map pmap;
-//    Vertex_Normal_Map nmap;
-//    double vlen;
-//};
-//
-//struct Top
-//{
-//    Top(VPMap pmap, VNMap nmap, double vlen)
-//        : pmap(pmap), nmap(nmap), vlen(vlen)
-//    {}
-//    void operator()(const vertex_descriptor& vin, const vertex_descriptor vout) const
-//    {
-//        put(pmap, vout, get(pmap, vout) + vlen * get(nmap, vin));
-//    }
-//    VPMap pmap;
-//    VNMap nmap;
-//    double vlen;
-//};
-
-
-
-
-
 
 // Property map associating a facet with an integer as id to an
 // element in a vector stored internally
